Fixed buggy window plane horizontal size
Was not considering that there is no x starting offset to check for
instead the remainder of the division must be considered.

This fixes Pacman but there is no reason why the sprites have
correlation to the amount of pixels drawn in the window.

There might be an out of bounds buffer access somewhere...?