Split up 16-Bit reads to prevent reading past buffer (+MBC1 fix)
Lode runner was executing a jump at 0x3FFE. The 2nd byte of the
target address was at 0x4000, so past the ROM0->ROM1 boundary.
The single 16-bit read was causing it to read the 0x4000th
byte of the ROM0 buffer, causing an out of bounds read.

Fixes HYPER LODE RUNNER crashes.

Also, changed MBC1/2 presence to be a global variable defined at
init time.